Combine strict equality with unspecified operator
