+++ src/parser/source/rules/noUnspecifiedOperator.ts
@@ -5,8 +5,8 @@ import { ErrorSeverity, ErrorType, type SourceError } from '../../../types'
 import type { Rule } from '../../types'
 
 const nonPermittedBinaryOperators: (es.BinaryOperator | es.LogicalOperator)[] = [
-  '==',
-  '!=',
+  // '==',
+  // '!=',
   // "**",
   '|',
   '^',
@@ -74,18 +74,45 @@ export class NoUnspecifiedOperatorError implements SourceError {
   }
 }
 
+export class StrictEqualityError implements SourceError {
+  public type = ErrorType.SYNTAX;
+  public severity = ErrorSeverity.ERROR;
+
+  constructor(public node: es.BinaryExpression) { }
+
+  get location() {
+    return this.node.loc ?? UNKNOWN_LOCATION
+  }
+
+  public explain() {
+    if (this.node.operator === '==') {
+      return 'Use === instead of ==.'
+    } else {
+      return 'Use !== instead of !=.'
+    }
+  }
+
+  public elaborate() {
+    return '== and != are not valid operators.'
+  }
+}
+
 const noUnspecifiedOperator: Rule<es.BinaryExpression | es.UnaryExpression | es.LogicalExpression> =
   {
     name: 'no-unspecified-operator',
     testSnippets: [
       ...permittedBinarySnippets,
       ...nonPermittedBinarySnippets,
-      ...nonPermittedUnarySnippets
+      ...nonPermittedUnarySnippets,
+      ['const x = 1 == 2;', 'Line 1: Use === instead of ==.'],
+      ['const x = 1 != 2;', 'Line 1: Use !== instead of !=.']
     ],
 
     checkers: {
       BinaryExpression(node: es.BinaryExpression) {
-        if (!permittedBinaryOperators.includes(node.operator)) {
+        if (node.operator === '==' || node.operator === '!=') { 
+          return [new StrictEqualityError(node)]
+        } else if (!permittedBinaryOperators.includes(node.operator)) {
           return [new NoUnspecifiedOperatorError(node)]
         } else {
           return []
@@ -109,3 +136,4 @@ const noUnspecifiedOperator: Rule<es.BinaryExpression | es.UnaryExpression | es.
   }
 
 export default noUnspecifiedOperator
+
